Moved references to be under the definitions package
