+++ python/src/aac/plugins/validators/reference_targets/_validate_reference_targets.py
@@ -1,7 +1,7 @@
 import logging
 
 from aac.lang.language_context import LanguageContext
-from aac.lang.references import is_reference_format_valid, get_reference_target_definitions
+from aac.lang.definitions.references import is_reference_format_valid, get_reference_target_definitions
 from aac.lang.definitions.definition import Definition
 from aac.lang.definitions.structure import get_substructures_by_type
 from aac.plugins.validators import ValidatorResult
